move files into lib; fix for linter
